Remove unnesseray hidden files created by R and Apple's Finder.
The .Rhistory is probably created by R and from its name it should not be included.
The .DS_Store files are created by Apple's Finder. It saves here some folder view settings, which however are personal and should not be saved in a respository anyway. Additionally, the best way for folder view settings is to set them, globally.