Shelf attempt to use NCBI ORFfinder instead get_orfs_or_cdss.py
Despite ORFfinder being a compiled binary, it does not seem to be faster than get_orfs_or_cdss.py. Additionally, it also has issues with sequence IDs and it does not save the descriptions of the entries. It limits the length of the ID to 50 characters, so that integrating the description into the ID by character replacement does not work either.