Make RAxML use an alignment file in fasta format instead phylip format
This removes the limitation of 50 characters per sequence ID and dot and hyphens are not replaced by underscores anymore.

This required to edit a couple LIT_1.1 files. The editing was done automatically with bash scripts. This way the sequence ID are identical in the *.tre and *.fas files, as they should be.